rework off chain code and test (runs, fails)
