+++ nitro-protocol/src/contract/multi-asset-holder.ts
@@ -64,126 +64,13 @@ export function convertAddressToBytes32(address: string): string {
  * @param allocation
  * @param indices
  */
-export function computeClaimEffectsAndInteractions(
-  initialHoldings: string,
+export function computeReclaimEffects(
   sourceAllocations: ExitFormat.Allocation[], // we must index this with a JS number that is less than 2**32 - 1
   targetAllocations: ExitFormat.Allocation[], // we must index this with a JS number that is less than 2**32 - 1
-  indexOfTargetInSource: number,
-  targetAllocationIndicesToPayout: number[]
-): {
-  newSourceAllocations: ExitFormat.Allocation[];
-  newTargetAllocations: ExitFormat.Allocation[];
-  exitAllocations: ExitFormat.Allocation[];
-  totalPayouts: string;
-} {
-  let totalPayouts = BigNumber.from(0);
-  let k = 0;
-
-  // copy allocations
-  const newSourceAllocations: ExitFormat.Allocation[] = [];
-  const newTargetAllocations: ExitFormat.Allocation[] = [];
-  const exitAllocations: ExitFormat.Allocation[] = [];
-  for (let i = 0; i < sourceAllocations.length; i++) {
-    newSourceAllocations.push({
-      destination: sourceAllocations[i].destination,
-      amount: sourceAllocations[i].amount,
-      metadata: sourceAllocations[i].metadata,
-      allocationType: sourceAllocations[i].allocationType,
-    });
-  }
-  for (let i = 0; i < targetAllocations.length; i++) {
-    newTargetAllocations.push({
-      destination: targetAllocations[i].destination,
-      amount: targetAllocations[i].amount,
-      metadata: targetAllocations[i].metadata,
-      allocationType: targetAllocations[i].allocationType,
-    });
-    exitAllocations.push({
-      destination: targetAllocations[i].destination,
-      amount: '0x00',
-      metadata: targetAllocations[i].metadata,
-      allocationType: targetAllocations[i].allocationType,
-    });
-  }
-
-  let sourceSurplus = BigNumber.from(initialHoldings);
-  for (
-    let sourceAllocationIndex = 0;
-    sourceAllocationIndex < indexOfTargetInSource;
-    sourceAllocationIndex++
-  ) {
-    if (BigNumber.from(sourceSurplus).isZero()) break;
-    const affordsForDestination = min(
-      BigNumber.from(sourceAllocations[sourceAllocationIndex].amount),
-      sourceSurplus
-    );
-    sourceSurplus = sourceSurplus.sub(affordsForDestination);
-  }
-
-  let targetSurplus = min(
-    sourceSurplus,
-    BigNumber.from(sourceAllocations[indexOfTargetInSource].amount)
-  );
-
-  if (sourceAllocations[indexOfTargetInSource].allocationType !== AllocationType.guarantee)
-    throw Error('not a guarantee allocation');
-
-  const guaranteeDestinations = decodeGuaranteeData(
-    sourceAllocations[indexOfTargetInSource].metadata
-  );
-
-  // for each guarantee destination
-  for (let j = 0; j < guaranteeDestinations.length; j++) {
-    if (targetSurplus.isZero()) break;
-    for (let i = 0; i < newTargetAllocations.length; i++) {
-      if (targetSurplus.isZero()) break;
-      // search for it in the allocation
-      if (
-        BigNumber.from(guaranteeDestinations[j]).eq(
-          BigNumber.from(newTargetAllocations[i].destination)
-        )
-      ) {
-        // if we find it, compute new amount
-        const affordsForDestination = min(
-          BigNumber.from(newTargetAllocations[i].amount),
-          targetSurplus
-        );
-        // decrease surplus by the current amount regardless of hitting a specified index
-        targetSurplus = targetSurplus.sub(affordsForDestination);
-        if (
-          targetAllocationIndicesToPayout.length === 0 ||
-          (k < targetAllocationIndicesToPayout.length && targetAllocationIndicesToPayout[k] === i)
-        ) {
-          // only if specified in supplied indices, or we if we are doing "all"
-          // reduce the current allocationItem.amount
-          newTargetAllocations[i].amount = BigNumber.from(newTargetAllocations[i].amount)
-            .sub(affordsForDestination)
-            .toHexString();
-          newSourceAllocations[indexOfTargetInSource].amount = BigNumber.from(
-            newSourceAllocations[indexOfTargetInSource].amount
-          )
-            .sub(affordsForDestination)
-            .toHexString();
-
-          // increase the relevant exit allocation
-          exitAllocations[i].amount = BigNumber.from(exitAllocations[i].amount)
-            .add(affordsForDestination)
-            .toHexString();
-          totalPayouts = totalPayouts.add(affordsForDestination);
-          // move on to the next supplied index
-          ++k;
-        }
-        break;
-      }
-    }
-  }
-
-  return {
-    newSourceAllocations,
-    newTargetAllocations,
-    exitAllocations,
-    totalPayouts: totalPayouts.toHexString(),
-  };
+  indexOfTargetInSource: number
+): ExitFormat.Allocation[] {
+  // TODO replace with proper implementation
+  return sourceAllocations;
 }
 
 /**
